image-processor-aws/requirements.txt
Flask>=2.0
boto3>=1.20
Pillow>=9.0
python-dotenv>=0.20
gunicorn>=20.0 # For deployment